requirements.txt
mysql-connector
flask
flask_restx
flask_cors
waitress
flask-jwt-extended
pandas